# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EthioMart E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user's orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/sellers:
    get:
      operationId: adminListSellers
      tags: [admin]
      summary: List all sellers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sellers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerListResponse"
  /admin/sellers/{id}/approve:
    put:
      operationId: approveSeller
      tags: [admin]
      summary: Approve or reject seller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveSellerRequest"
      responses:
        "200":
          description: Seller status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
...
  /admin/products/{id}/approve:
    put:
      operationId: adminApproveProduct
      tags: [admin]
      summary: Approve or reject product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveProductRequest"
      responses:
        "200":
          description: Product status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Get platform analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformAnalytics"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required: [email, password, firstName, lastName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, seller, admin]
        avatar:
          type: string
        address:
          type: string
        city:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, firstName, lastName, role]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        discount:
          type: integer
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        sold:
          type: integer
        featured:
          type: boolean
        status:
          type: string
          enum: [active, pending, rejected, inactive]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, price, stock, categoryId, sellerId, rating, reviewCount, status]
...
    AddToCartRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: string
          enum: [telebirr, cbe_birr, cash_on_delivery]
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        trackingNumber:
          type: string
        estimatedDelivery:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orderNumber, userId, items, status, paymentStatus, subtotal, shipping, total, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required: [id, productId, quantity, price, subtotal]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, totalPages]
    CreateOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AddToCartRequest"
        shippingAddress:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
          enum: [telebirr, cbe_birr, cash_on_delivery]
        couponCode:
          type: string
      required: [items, shippingAddress, paymentMethod]
...
    Address:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        region:
          type: string
        country:
          type: string
          default: Ethiopia
        landmark:
          type: string
      required: [fullName, phone, street, city, region]
...
    Seller:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        storeName:
          type: string
        storeDescription:
          type: string
        logo:
          type: string
        category:
          type: string
        city:
          type: string
        rating:
          type: number
        totalSales:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, storeName, status]
...
    SellerListResponse:
      type: object
      properties:
        sellers:
          type: array
          items:
            $ref: "#/components/schemas/Seller"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [sellers, total, page, totalPages]
...
    PlatformAnalytics:
      type: object
      properties:
        revenue:
          type: number
        orders:
          type: integer
        users:
          type: integer
        sellers:
          type: integer
        revenueChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        userGrowthChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        categoryRevenue:
          type: array
          items:
            $ref: "#/components/schemas/CategoryStat"
        paymentMethods:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
      required: [revenue, orders, users, sellers]
    ChartDataPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: string
      required: [date, value]
...
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [status, count]
    CategoryStat:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        count:
          type: integer
        revenue:
          type: number
      required: [id, name, count]
    ApproveSellerRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected, suspended]
        reason:
          type: string
      required: [status]
    ApproveProductRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, rejected]
        reason:
          type: string
      required: [status]